# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "AppController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Đăng nhập thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Sai thông tin đăng nhập"
          }
        },
        "summary": "Đăng nhập",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Đăng ký thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ hoặc username đã tồn tại"
          }
        },
        "summary": "Đăng ký tài khoản",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/users/me": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/all": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "users"
        ]
      }
    }
  },
  "info": {
    "title": "KTPM API Documentation",
    "description": "API Documentation for KTPM project",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "Authentication",
      "description": "API liên quan đến xác thực và quản lý thiết bị"
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "john_doe"
          },
          "password": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "admin@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "1"
          },
          "phone": {
            "type": "string",
            "example": "0987654321"
          },
          "email": {
            "type": "string",
            "example": "admin@example.com"
          },
          "dateOfBirth": {
            "type": "string",
            "example": "2000-01-01"
          },
          "name": {
            "type": "string",
            "example": "Nguyễn Văn A"
          },
          "role": {
            "type": "string",
            "example": "Owner"
          }
        },
        "required": [
          "username",
          "password",
          "phone",
          "email",
          "dateOfBirth",
          "name",
          "role"
        ]
      }
    }
  }
}